revert header and devcontainer cleanup
